My first data analysis project
This my first guided project I did with dataquest.io. I had to implement things that I learned in earlier lessons like for loops, indexes, and frequency tables with two list of lists(Without Dictionaries). 

The objective was to clean the data, collect information about helicopter prison escape attempts, and to display that data with a frequency chart.